Добавлено модульное тестирование клиента и сервера

+++ server/src/main/java/ru/hj77/server/dto/ClientDTO.java
@@ -7,10 +7,11 @@ import java.util.List;
 
 @Value
 public class ClientDTO {
-    int id_client;
+    Long id_client;
     String name;
     String surname;
     String patronymic;
     Date date_of_birth;
     List<CardDTO> cardsDTOList;
+
 }
